refactor(frontend): move widget to app-components
