+++ frontend/src/layout/index.tsx
@@ -7,7 +7,10 @@
  */
 
 import { BoxProps, Grid } from "@mui/material";
+import { usePathname } from "next/navigation";
 import { useState } from "react";
+
+import { ChatWidget } from "@/app-components/widget/ChatWidget";
 
 import { Content } from "./content";
 import { Header } from "./Header";
@@ -26,6 +29,7 @@ export const Layout: React.FC<LayoutProps> = ({
   sxContent,
   ...rest
 }) => {
+  const pathname = usePathname();
   const [isSideBarOpen, setIsSideBarOpen] = useState(false);
 
   return (
@@ -42,6 +46,7 @@ export const Layout: React.FC<LayoutProps> = ({
       <Content sx={sxContent} {...rest}>
         {children}
       </Content>
+      <ChatWidget pathname={pathname} />
     </Grid>
   );
 };
